backend/requirements.txt
# Core framework
Django>=5.0,<5.1
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
django-filter>=24.3

# Database
psycopg[binary]>=3.2

# Async & realtime
channels>=4.0
channels-redis>=4.1
celery>=5.4
redis>=5.0
django-celery-beat>=2.6
django-celery-results>=2.5

# Utilities
python-dotenv>=1.0
drf-spectacular>=0.27
Pillow>=10.4
PyPDF2>=3.0
pytesseract>=0.3
pdf2image>=1.17
python-docx>=1.1

# Testing & quality
pytest>=8.0
pytest-django>=4.8
factory-boy>=3.3
faker>=24.0
black>=24.8
isort>=5.13
flake8>=7.1

# Deployment
gunicorn>=21.2

